refactor(chat): extract chat_manager/config_api.rs (#292)
Move update_chat_config, enable_worktree, attach_worktree,
worktree_offers_for_chat, dismiss_worktree_offer,
accept_worktree_offer, disable_worktree, fork_to_worktree, and
remove_project off the ChatManager facade — pure move, no behavior
change.
